Updated openstack/openstack Project: openstack-infra/project-config  7ed9c6ee2c09b0602e04ae0d11db221226b677fd
Expand jobs for releases repository

Add a list-changes job to show the commits expected to be part of the
new release(s) proposed.

Add pep8 checking, since we have more python code now. There are no unit
tests for the code, so we still don't need py27.

Add the validate job to the gate queue (it was already in the check
queue) to ensure we don't have issues with competing changes to the same
repository merging somehow.

Change-Id: I749ad1afb253f6c822e2cd4c454ef40ff24f6ed7
Depends-On: If8a77cf24e8a0ac03dd065c5813960696730b864