Updated openstack/openstack Project: openstack/ceilometer  b0c8b47ce7d5e19a62aa889b316a80d9ab463b07
Close and dispose test database setup connections

Without explicitly closing and disposing the connections can linger in a
non-deterministic fashion and cause intermittent "too many connection"
errors in the tests.

This is effectively a "backport" of change
I8badf2a3094f2d27d381bd178c6d579f2a5c5ffe in aodh, made here for
safety despite not seeing the problem regularly in the ceilometer
tests.

The -N 100 is removed from the setup-test-env so that we are using
whatever the postgresql default is, meaning that these tests will be
better at exposing errors that users may see.

Change-Id: I694d118ff8dc2b5c8be633121fa1037130fbe874