Updated openstack/openstack Project: openstack/castellan  4088221f280438c7a91831cc53371d227221ad58
Remove copy_key operation

This change removes the copy_key operation from the key manager. The
copy_key operation isn't ideal because few key managers support such
an operation natively. Lack of native support requires the key to be
retrieved and stored in separate steps, which increases the handling
of the key material.

It would be relatively trivial to add this operation back to the
key manager interface at a future point. Once Castellan becomes
widely used by other projects, removing this operation will not be
possible, as it would be a backward-incompatible change.

Change-Id: I1a1dfdb4d4268319f9277fc639027819e70d4a8b