Updated openstack/openstack Project: openstack-infra/project-config  360449421546ac5bfd407ddcd6464b438d53defd
puppet: modify beaker/upgrade job to test kilo to liberty

We first created the beaker-upgrade job to test if master patches did
not break HEAD^1.
So we used to checkout HEAD^1 of the patchset, run beaker, checkout HEAD
and run beaker again. It was for testing and getting early feedback of
job behavior.

Now, we would like to move forward and test real upgrade between
stable/kilo and master (which is supposed to be Liberty soon, once we
have packages).
The workflow stays the same, but instead of checkout-ing HEAD^1, we git
checkout stable/kilo, run beaker, checkout the patchset and run beaker
again. Puppet should take care of upgrading OpenStack and the job will
let us know if services are still running if resources managed by Puppet
survived.

Change-Id: Ifc88787b2d7ed469f530cc7cc4c5565064b89602