Updated openstack/openstack Project: openstack/cinder  dde1b685fba29a2c60188de79e8539ffc991b07f
Fix cleanup_temp_volume_snapshots for missing vol

The cleanup_temp_volume_snapshots method raises an exception that
causes the backup service to exit if there is a backup without a
corresponding volume.

This commit catches these exceptions so that the backup service is
not prevented from starting in this circumstance.

Change-Id: Ia1aac2fe78048df5fce595d5de181cb4930ea78d
Closes-bug: 1478599