Updated openstack/openstack Project: openstack/swift  8a8c3bc39aa91039be6d7f1fea010c4b25d5e1be
EC tests should use ECDiskFileManager

Some EC tests in proxy/test_server.py use the replication
DiskFileManager to open diskfiles. That works because it
turns out that DiskFileManager is robust to the frag index
being stitched onto the end of the timestamp in filenames,
but it would be better to test using the ECDiskFileManager
for EC diskfiles.

Also, drive-by deletion of some duplicated lines.

Change-Id: Ie3adb607d8c34a469df7d0f6f542ed5ba9a2dac0