Updated openstack/openstack Project: openstack-infra/puppet-graphite  c7bec6c4b5ded8c37733072470b6cde73c00551c
Make username/email/password params required

Values for the graphite_admin_username, graphite_admin_email, and
graphite_admin_password parameters must be provided by the user. The
empty string is not a sane default and will cause broken deployments if
used. This change makes these parameters required. Infra is providing
these parameters in system-config so this change will not break Infra.
If downstream users are not providing these parameters their
deployments will not be functional, so this change should be safe for
users already using this class correctly.

Change-Id: I7aed54bb09e54dc8c0827188adf971b76649f94d