Updated openstack/openstack Project: openstack/heat  0348e8571e9e8d9370ebd47018fa3674fb11e3fb
Convergence: Handle InvalidTemplateAttribute in worker

All attributes are retrieved (including ones referenced in the output).
In our functional test there is a bug where a nested stack did not
have an output that was referenced from the output. This didn't effect
the stack creation as:
- outputs are not normally required in the stack creation
- if the output is used by a TemplateResource it *may* be used

The way outputs are used currently is that they do not fail on
the first output that doesn't work as it is useful to get the ones
that do work. To help with this errors are placed alongside the
value (in a key "error_msg").

This is somewhat problematic in convergence as we now *require* all
attributes (including those based off of outputs).

Note: if you have an outer resource referencing a non-existent template
resource output it *will* fail, but when the outer stack's output references
the inner stack's output then it is not validated.

Change-Id: Id07c617a19eae56543f92ee21aea58cd38fa3606