Updated openstack/openstack Project: openstack/horizon  f08f3fa3701ad92b7fee219bf885fa13dc647152
Create image directly in 'copy_from' and 'location' case

The Glance API will return immediately with a response. We need this
response to give a errror message back to the user in case of invalid
URL.

Any download operation will be executed in the background by the
Glance API server and there is no need for executing this step in a
separate thread.

Keep current logic for updating image with posted data in a background thread
One can argue that this too is uneccessary, but I can see that it may save
time for large images or when bandwidth between Horizon and Glance is limited.

Change-Id: I2ec9f714f9cfd6ca2271750133d6fcac29b94f35
Partial-Bug: 1438975