Updated openstack/openstack Project: openstack/sahara-image-elements  c3929c575bac6f45c36970dceb074ac692ff3ad0
Deprecate Spark 0.x and 1.0.x images

This change request adds deprecation warnings when generating
images containing Spark versions that have been deprecated in the
Sahara Spark plugin.
The idea is also to deprecate all Spark versions that depend on
the hadoop-cdh element, so that it can be removed later on.

Change-Id: Iefed31c8f22a40eb4e3eed835d7bb10f4fa20750
Implements: blueprint deprecate-spark-100