Updated openstack/openstack Project: openstack-infra/shade  3e8f419c70055108f7ac8ba3f079ec88fdec7500
Fix for swift servers older than 1.11.0

Discoverable capabilities was not added to swift until 1.11.0. Some
providers are running an older version. These will return a swift
exception with an HTTP status of 412 because the URL (/info) is bad.

This change sets a default that is half the Swift default max file size.
The result is also cached as it isn't expected to change.

A separate, future change will take advantage of a newer os-client-config
which will allow setting the default value per-vendor profile.

Change-Id: I80b14cfae42fc160f42b44e4923472d14a4b9057