Updated openstack/openstack Project: openstack/nova  6098f4727ccbd8a8dd84cee4654ba8308ff752bf
libvirt: implement get_device_name_for_instance

We implement the method above with the purpose of using the
nova.libvirt.blockinfo primitives (which match how libvirt itself works),
to decide on which device name to use for an instance.

With this change - we will be using the libvirt local logic to decide on
the volume device name instead of relying on
nova.compute.utils.get_device_name_for_instance() which will do wrong
things in a number of cases.

Closes-bug: 1452224
Related-bug: 1231874

Change-Id: I33d7363c018fc4e5e9f190486d5556363852badb