Updated openstack/openstack Project: openstack/manila  26aa8773530005524466f61af0667f9c53d11e47
Eventlet green threads not released back to pool

Presently, the wsgi server allows persist connections hence even after the
response is sent to the client, it doesn't close the client socket connection.
Because of this problem, the green thread is not released back to the pool.

In order to close the client socket connection explicitly after the
response is sent and read successfully by the client, you simply have to
set keepalive to False when you create a wsgi server.

DocImpact:
Added wsgi_keep_alive option (default=True).
In order to maintain the backward compatibility, setting wsgi_keep_alive
as True by default. Recommended is set it to False.

This is port of Cinder change - [1]

[1] Ic57b2aceb136e8626388cfe4df72b2f47cb0661c

SecurityImpact
Closes-Bug: #1361360

Change-Id: If9241e6f6ba10592a64ca312cb479e8cea929913