Updated openstack/openstack Project: openstack-infra/elastic-recheck  b0fddde7101784a6a36d9eaec680fe1e428b3b27
Remove old queries for "fixed" bugs (07/23/15)

Removes queries that haven't hit in over 10 days
on bugs that are either fixed, or are incomplete/invalid
to the point where they aren't issues anymore -
or at least not seen in a long time.  If they come back,
we can write a new query for a new bug.

Change-Id: I560a7e2b886f05c45a6cbf061deec1a7d8a258ac