Updated openstack/openstack Project: openstack/manila  3af8f5f6dcc04359cef6274e77c2c3fd1e397d4f
GlusterFS: fix retrieval of management address of GlusterFS volumes

We need the so-called management address of a GlusterFS
volume (ssh address of its management node) at a few places
for bookkeping reasons (certain data like GlusterFS version
is stored in management address-keyed dicts).

So far it was available only through ad-hoc mangling of
various attributes of a GlusterManager instance that
represents the volume. Now we make it available
directly from GlusterManager as an attribute.

Also fix an occurrence of ad-hoc mangling going wrong,
ie. where a wrongly constructed address is attempted
to be used as dict key.

Change-Id: Ic5a96bc99943dda3592372512916257d53f61b80
Closes-Bug: #1476710