Updated openstack/openstack Project: openstack/tempest  74647862be3b87687b27628d64da733ce8ad488c
Update smaple config for new oslo.config release

The 2.1.0 release of oslo.config changed another default config value.
This causes the sample config job to break because the config options
not defined in the tempest repo to change. This commit fixes this for
the time being. However, as a long term fix the job must be changed to
only generate a sample config for in tree options, and leverage tempest
init to generate a full config for tempest plus the oslo libs.

Change-Id: I053d5d464d0b5eb5a7902338efa3cb25d2fd7f74