Updated openstack/openstack Project: openstack/oslo.versionedobjects  91184e1047d07c8fbeee2648ec7a929858b58b89
Fix ComparableVersionedObject in python 3.4

There is a difference when __eq__ is called and not called
in python 2.7 and 3.4.  In python 3.4, when an object is
compared to something else, e.g. obj != None, the __eq__
is called to do the comparison, whereas in python 2.7,
it does not.  The following patch allows
ComparableVersionedObject to be used in python 3.4 by
checking if the object has obj_to_primitive() as an
attribute.

Change-Id: I9b989bb07505842651bc73c2ccc1552b14307a68
Closes-Bug: #1469864