Updated openstack/openstack Project: openstack/nova  1a532be8f42c0cb3debd550cf8d8f46e869386d6
Refactor test_servers to use instance objects

This does a major refactor of code in test_servers to use instance
objects and generally mock things at the compute/api layer instead
of all the way down in the database layer.

It removes a few tests and conditions that were just testing DB
semantics from the highest levels.

This will be needed when we drop the flavor compatibility code from
the instance object, as these tests were mocking out the DB layer
with old-style instance structures.

Change-Id: I42ffac6b323a809ee80a1e8dedf729de46106b47