Updated openstack/openstack Project: openstack-infra/project-config  10e1c4206e919e7bbea198419a444a8977b5bfe7
Disabled neutron-dvr job for stable branches

It was enabled lately to vote, though stable branches do not contain the needed
patches to make it stable.

It's not expected the branch will receive the needed backports to fix it
in Kilo, so the job is doomed to be useless and is better to be disabled
for those branches instead of heating cpu in gate.

While at it, cleaned up an old entry to avoid the job for Icehouse since
the branch is already removed.

Change-Id: I3ba2b38ff2b82dd971ecd9500c235f8dd8320aa1