Updated openstack/openstack Project: openstack/cinder  2c747c0a04fcd58945de754ebc902f23e000a3a9
Adding NFS support to the GPFS Driver

This change helps create cinder volumes on a GPFS cluster,
when the cinder services are not running directly on one of
the GPFS nodes.

The NFS export created on GPFS cluster will be locally mounted
on a node running the cinder service. The cinder volume service
will communicate with GPFS nodes over SSH to execute CLIs remotely.

This change also ports the GPFSDriver to ABC driver model.

Implements : blueprint ibm-gpfs-nfs-support

DocImpact

Needs an update in gpfs driver documentation, as this patch
adds an NFS based access for GPFS cinder driver.

Change-Id: I40e5098482d48552170c34e74be0c9e2f544fef5