Updated openstack/openstack Project: openstack/sahara-specs  457b6b6107a2af4a7f101a6b666c0ca3919e48e5
API to mount manila shares to clusters

Manually mounting shares to every node in a large cluster would be a tedious
and error-prone process. Auto-mounting shares that are requested for use in
either the data source or binary storage case might be feasible for some use
cases. However, outside of our (optional) EDP interface this functionality
would never be usable. As such, it is best to provide the user an API for
mounting of shares onto Sahara clusters.

Change-Id: I980c7514a11bb8755f902312bc2d2c118619cce0
Addresses: blueprint mount-share-api