Updated openstack/openstack Project: openstack/barbican  83d36f46412f416981473abcf7e8d00fdd2b80dc
Revert backwards incompatible paste change

A breaking change to the paste.ini was introduced in git commit
3b20d84312b6870557a5b1db346de75e928ddbe8 where the main app was
renamed from "apiapp" to "apiapp_v1".

This was an unnecessary change that would force deployers to rewrite
their configuration files, which may include changes to configuration
management systems.

This change reverts the incompatible name change so that operators are
able to upgrade from Kilo to Liberty witouth having to rewrite their
paste file.

Change-Id: If3dc7060ae5ea6d427c3a1e9989fa1e17a469db7