Updated openstack/openstack Project: openstack/requirements  0f6c75d5cfc64a2a19a3d3b48d1a4e563c5c2c22
List 'singledispatch' in global-requirements

The 'singledispatch' backport is useful for people to be able
to use (as its included by default in functools in 3.4 and newer)
and at least taskflow would like to use it so that the functionality
works the same in 2.7 and 3.4 (and newer).

It is already listed in upper-constraints but not in
global-requirements but this adjusts that, allowing it to be used
by projects and libraries in openstack.

Change-Id: I3c8ff515c29d8156ad114ba6dfb36bcfc68689e0