Updated openstack/openstack Project: openstack/heat  0cafeac1d90de5b655e6df54fe8451a467485f21
Refactor and correct purge_deleted

This method had not been updated in a while
to account for changes in the db schema. Also,
it was refactored to use bulk delete statements
rather than iteration as the iteration approach
ran quite a long time and consumed quite a bit
of resources.

Change-Id: Ia49e801d3c244a1fabb8441e7eaa22b88e5fcb56
Closes-Bug: #1439518