Updated openstack/openstack Project: openstack/heat  480e63b8a86e8e2eec56c528b1e977d8eb551352
Correct stack identity when delete nested

1. Update an aws asg with invalid 'VPCZoneIdentifier',
the stack update failed.
2. Update again, the stack is always in 'UPDATE_IN_PROGRESS'.

We remove the backup resources first when updating,
but the backup of stack_resource can't be deleted,
due we set an incorrect name of stack_identity before
rpc calling 'delete_stack', due the name is not match the
nested_stack's name.

Change-Id: I89966d0124607c233ac9a82f1ce14f4f2deb3f9b
Closes-Bug: #1478480