Updated openstack/openstack Project: openstack/requirements  9bbc6ddce3ab71a234010dd1e95acb82a9ab6bd0
Add dnspython3 to global-requirements.txt

dnspython3 is the official py3 compatible release of
dnspython, maintained and released by the same authors.

Designate makes heavy use of dnspython, and is in the
process of adding py34 support.

Additionally, add environment markers for dnspython and
dnspython3 to ensure the correct versions are used.

Change-Id: Ib652e68c21fbe50fc849eacbde5d1c856aa4f254