Updated openstack/openstack Project: openstack/trove  ee56b6a28e9484edc969b8c4407d0277c2f2b9fa
Add support for extended_properties for clusters

During cluster creation, we are now able to pass in custom
properties specific to a datastore as part of a data-bag
named 'extended_properties'. This allows us to get these
properties to the appropriate clustering strategy which can
then use them to implement datastore specific actions.

Change-Id: If24f5bf6181bf1502042ac51868a161a6a037bf2