Updated openstack/openstack Project: openstack/trove  04621808e3f7e84392df070c67fb7ec359e59372
Cluster instances could falsely report 'ready'

The check used by all cluster strategies is wrong. It allows for both
RUNNING and BUILD_PENDING. All cluster strategies use BUILD_PENDING to
indicate that the instances are cluster-prepare ready. RUNNING is bad as
it can be reported before the instance's prepare even starts, causing
the cluster-prepare to start too early and fail.

The fix is to remove RUNNING from the list of cluster-ready states. It
has been verified that all datastores  set instances to BUILD_PENDING
when used for clusters.

Change-Id: I048e2b76fda59a25928c6c6d3751235df12d7369
Closes-bug: #1478603