Updated openstack/openstack Project: openstack/nova  a39674bf13d3eed32a2d67bfb13ddc28a5da0b28
tweak graphviz formatting for readability

The "group" tag on nodes can be used to create vertical alignment, all
nodes in a group which are connected will end up in one column. This
makes reading the flow chart a bit easier.

Add a " " before the "no" edges otherwise they are left justified
completely to the vertical edge, and make it kind of gorpy to read.

Change-Id: Ife11e905ea6fdd97e1459fbbc001324eb946f9e4