Updated openstack/openstack Project: openstack/tripleo-heat-templates  5a6e934b4b8d6fab7b899c368fb82321e1a0cbe3
RHEL registration shouldn't use attach with activation key

In the case of using portal registration with an
activation key, the RHEL registration script is still
executing a `subscription-manager attach` command. This
should not happen if an activation key is provided. This
is because an activation key already provides the
subscriptions to attach.

Change-Id: I2907bede28a9b7bef71cedeea69c876eb4949df0