Updated openstack/openstack Project: openstack/nova  4eb2d2f707bce3480de5fe44446c90e1d2ff679b
virt: convert hardware module to use nova.objects.ImageMeta

Convert the APIs in the nova.virt.hardware module to expect to
receive an instance of nova.objects.ImageMeta instead of a
dict.

Update the virt drivers to convert the dict into an ImageMeta
module, as a stepping stone, until they to are converted to
use ImageMeta throughout.

Change-Id: Iac034d864cd0265759aaa7c8049fafa276763cc0