Updated openstack/openstack Project: openstack/nova  8eabbd64e52fcf330e8defd76f32cf7d0f8a9690
libvirt: handle rescue flag first in blockinfo.get_disk_mapping

If rescue is True, we don't need to get the flavor object from the
instance (which could lazy load it from the database) or figure out the
pre-assigned device names for the block_device_info list.

This just re-arranges the order of calls to be more efficient.

The related unit test is updated to show that get_disk_mapping works if
rescue=True and there is no Flavor object in the instance passed in.

Change-Id: I5ac7e4d3129c67db4efde0dda71c41b747952263