Updated openstack/openstack Project: openstack/nova  c2e70f71ce9e1f5deae99632485aedfbb5bd04e2
libvirt: update docstring in blockinfo module for disk_info

The docstring was listing disk_info as a tuple but it's a dict in the
code and it wasn't pointing out that boot_index is an optional field.

Long-term we should probably convert disk_info into an object (doesn't
need to be versioned, but would be cleaner as an object). I started
doing this with a namedtuple but the format and boot_index fields being
optional makes namedtuple a bit messy.

Change-Id: I6721cad9245e8722481c4c63023bb8b7193317ac