Updated openstack/openstack Project: openstack/nova  31396bc1b1b891218e61eaf1d53aa2ddd104c96e
libvirt: move the LibvirtQuobyteVolumeDriver into the quobyte module

The setup in LibvirtVolumeTestCase is moved to a common base test
class so we don't run tests multiple times by extending the
LibvirtVolumeTestCase class in the type-specific volume driver
tests. Helper methods are moved as needed for the tests.

Part of blueprint consolidate-libvirt-fs-volume-drivers

Change-Id: I0a6896e7d1de58b4550f18819113a38799a1fa12