Updated openstack/openstack Project: openstack-infra/project-config  7d266b85425ba0a3fd194a4cb0be0bc8fa83545a
Run full multinode tests against new dib images

We are checking that the new dib built images that use a minimal base
and glean to init on boot are capable of running our tests. We should be
testing the full test suite and not the smoke tests. Switch the job over
to do this.

Change-Id: Ic3984df20611691ecd0328f5d678d5b9fab5a677