Updated openstack/openstack Project: openstack-infra/project-config  b4d206c36d6bba8ed1a44dbbf59c4812b9c6e7a6
Make the Heat Convergence Gate job non voting

Now that this gate job is reasonably stable, hardly a test or two
failing, it would be good to get some more test results and log files
to help debug the last few bugs that remain. So, make it run on the
check gate but keep it non-voting still.

Change-Id: I41b7af5ceacef89419cc87c28e8e964cfba08cc1