Updated openstack/openstack Project: openstack/manila  200b36391e636d1e8f3c3266454aeb3a14f7dcc1
Listen on :: instead of 0.0.0.0 by default

If Manila API runs on a system with IPv6, it listens on IPv4 by default,
but clients try IPv6 first. The resulting retry is unnecessary. On Linux,
it is easy to listen on both address families simply by opening an IPv6
listening socket, so this patch makes use of that feature.

The listening address may be set explicitly as before. Only the default
is changed by this patch.

Closes-Bug: #1338825
Change-Id: I45dd3c94b332d14a49c78f480cf1f6f69749acf3