Updated openstack/openstack Project: openstack-infra/irc-meetings  c68af8eac7b46a1d50056a8ce72d68da594a8f52
Update Searchlight Team Meeting

The meeting was added prior to Searchlight being accepted as a new project.
This updates the text to be "Searchlight" instead of just search and
also provides a link for adding agenda items.  The time and IRC room
does not change.

Change-Id: I0302aab1a13c971aa16eefee83566e01aa3e41db
See: https://review.openstack.org/#/c/188014