Updated openstack/openstack Project: openstack/tooz  d01b7d7018b510d23a27249a2c8330e4936f1972
Accept blocking argument in lock's context manager

This patch adds the posibility of passing the blocking argument to the
context manager to be used when acquiring the lock.

Previous behavior will be preserved, so now we will be able to use the
context manager in any of the following ways:

    with lock(10.0):
    with lock(False):
    with lock():
    with lock:

The last 2 are equivalent.

Change-Id: I6d2f486ac45078ec4ce77de52b7997f2122d4dc8