Updated openstack/openstack Project: openstack/nova  3fa3ea7def105dfb45563b927f83ee974f7d3525
Add DictOfListOfStrings type of field

For Scheduler hints, sometimes some hints can be either strings or list of strings.
In that case, we prefer to accept all of them as list of strings and provide a way
to provide a DictOfListOfStrings type of field for the future scheduler_hints field
of the RequestSpec object.

NOTE: this type is also proposed in o.vo by https://review.openstack.org/205110

Partially-Implements: blueprint request-spec-object

Change-Id: Ibf6e9b83dcd676bab076821f8eca335974d49f06