Updated openstack/openstack Project: openstack-infra/devstack-gate  420a213ddf23f04504fcdb868966a5fe10046696
make package install safer

The following makes apt-get install quite a bit safer by retrying the
apt-get update should it fail using a common pattern copied from
devstack.

It also keeps track of whether or not we've run the apt-get update
during this run, which means we don't it more than once for no reason.

For the specific case of the python yaml module we also check to see
if it's installed first (which it often is) to avoid doing network
actions when not needed.

This should make the

Related-Bug: #1400905

show up less often

Change-Id: I77159da4e33f4255bd8a5d7d8cd3eaae3a8a59a4