Updated openstack/openstack Project: openstack/barbican  4cd678cf0c5b803634cef7e44f39838df196272c
Flatten exceptions used in policy tests

The policy tests are meant to check if certain roles are properly
enforced by our policy engine. To do this we use the standard
controllers and force an exception in the actual code, just after we
have actually passed the policy.

The code was actually doing what it intended to do, in an unintended
way. By passing the "external_project_id" to the
on_[get|post|delete|put] it was finally getting a TypeError when it
would try to execute the actual code. This was unexpected and
confusing.

The intent of this CR is to get rid of the aforementioned behaviour and
instead force an IOError whenever possible (either by accessing certain
functions, or when trying to access the body of the request).

Change-Id: I98d168c1fe249c1fc5e5182cac43f63a07fa2e3a