Updated openstack/openstack Project: openstack/tuskar-ui  3742ba77a7765c51f2a431c85c482c960a9a7337
Do not sync parameters across roles

Wholesale syncing of the parameters across roles doesn't work. There are
some parameters that can be different depending on the role. For one,
NeutronPublicInterface might be different depending on the node's role.
This patch comments out the sync. Future might involve creating a
white/blacklist for which params should/should not be synced.

Change-Id: I66be8126e0160aaa0af2f49b7d5fce705e720f52