Updated openstack/openstack Project: openstack-infra/project-config  764befa2b2e9cbb73d13876b0ad7b3e1213c3374
Write post job logs to apache and swift

Currently we don't get generated parent dir indexes for swift uploaded
logs. This is mostly ok for check/gate jobs as we address the log dirs
directly in the path supplied to gerrit and presented to users. For post
jobs we do not do this so rely on something to provide a proper index.
Go back to apache uploaded logs in addition to swift uploaded logs so
that these files will be indexed until we can set up indexing with
swift.

Change-Id: I7f688834e0d3ad91b81dcbde7643a4b5c208748e