Updated openstack/openstack Project: openstack/cinder-specs  ea94724102d1f3d592b96f46fff8b1b15bb0bab2
Add ability for Cinder backend to report discard/unmap/trim

Add the ability for Cinder backends to report support for discard
capability. If a backend driver supports discard it can add a
"discard": True entry in the connection properties returned from
initialize_connection. This will be used by Nova to ensure any
attached volumes have support for this functionality. A separate
Nova patch is in the works as well.

Change-Id: I3335fd93b9603506941ff456e62d6b0757d9e602
Implements: bp cinder-backend-report-discard