Updated openstack/openstack Project: openstack/swift  8cb6d76335c1c4edb8ea8260663058362548e59e
Swift documentation for first-time contributors.

In this change, there are proposed improvement to Swift documentation aimed at
first-time contributors. They include a simplification of the Getting
Started page and a new page with some basic instructions/commands that the
first-time contributor should know. In addition, it shows some common errors
that the first-time contributor may find when executing git rebase and
information on how to track your changes.

Change-Id: I704202955093736b2f3b4102a649690a0392c6b0