Updated openstack/openstack Project: openstack/nova  c44405514bf0304011e0f2cdf53020562e276069
VMware: create common object for limits, reservations and shares

Currently we are supporting cpu limits, reservations and shares.
The same are also supported for disks, memory and vnics.

This patch creates a common data structure to be used to support all
of the above.

This is part of the blueprint vmware-limits

Change-Id: I0abf816a6e3b29f10f7380a5865d30e887c09766