Updated openstack/openstack Project: openstack/nova  b95044862af0e731e586433bd3c3a6585ff98fbb
VMware: add resource limits for disk

The flavor extra specs for quotas has been extended to support
 quota:disk_io_limit - The utilization of a virtual machine will not exceed
                       this limit. The unit is number of I/O per second.
 quota:disk_io_reservation -  Reservation control is used to provide guaranteed
                              allocation in terms of IOPS
 quota:disk_io_shares_level - the allocation level. This can be 'custom', 'high'
                              'normal' or 'low'.
 quota:disk_io_shares_share - in the event that 'custom' is used, this is the
                              number of shares.
This enables the VMware driver to enable disk allocations
for the running instance.

Please see http://goo.gl/6sHwIA for a detailed explanation of the above.

DocImpact

The is part of the blueprint vmware-limits

Change-Id: I6797ad3442895cd73f240590286b4ae13986291f