Updated openstack/openstack Project: openstack/nova  026cc5a86bd115f53e1522247ee7b4192e5726b8
libvirt: Add macvtap as virtual interface (vif) type to Nova's libvirt driver

This is required in order to attach libvirt managed KVM guests via macvtap
to the hosts network. It will be exploited by a new macvtap ml2 plugin and
agent. The place for this new support is currently discussed [1].

[1] https://review.openstack.org/#/c/195907/

Implements blueprint libvirt-macvtap-vif

Change-Id: I06047913b66cb8d0eaa0a5d08d0cfb13aadd260a