Updated openstack/openstack Project: openstack/openstack-manuals  0eea064660cea1f49a1b9773eaa5b776415a85f7
Update link to baremetal provisioning

The link to baremetal under the hypervisors section of the Cloud Administrator Guide
points to a depreciated page for nova baremetal, instructing users to go to the Ironic
baremetal provisioning page instead.

The fix updates the baremetal link within the Cloud Admin Guide rst 
and xml files to go directly to the Ironic page.

Change-Id: Ia8607a84832b1f53ad54d38e75b1687d579da0eb
Closes-Bug: #1478628