Updated openstack/openstack Project: openstack/nova  e0c27948d8662eb69985d69e6f30aacd183b81fb
return more details on assertJsonEqual fail

assertJsonEqual does a recursive traverse of a data structure, using
assertEqual strategically along the way. The problem with this is that
it can lead to really cryptic messages like

  Failed '3' != '4'

When some very deep list was the wrong number of elements. It doesn't
tell you which list was the wrong number of elements, and so it's
completely opaque to get to the bottom of. Especially if this doesn't
happen on every run, just some.

This creates a version which catches an inner match issue, and
constructs a new failure which includes information about the 2 data
structures being compared at the top level. This should help debugging
racey unit test in the gate like
test_describe_instances_with_filters_tags.

Related-Bug: #1479780

Change-Id: I6f6e961a3c63f9f86fe9b21ce6c16a2e634e9ce5